restore(input): connection nav + copy on Details, filter typing across tabs
Two behaviors lost during the Phase 5 migration are restored:

1. Navigation (j/k/Up/Down, g/G, PgUp/PgDn, Ctrl+B/F) and 'c'
   (copy remote address) now work on the Details tab again. They
   flip which connection has selection focus — on Details that's
   also what drives the displayed record, so users can page
   through connections without bouncing back to Overview.

   Extracted to ui::try_handle_connection_nav so OverviewTab and
   DetailsTab both call it. OverviewTab's match shrinks by ~50
   lines; DetailsTab::handle_key is the new four-line delegation.

2. Filter typing keeps working when the user switches to Details
   while filter mode is active. ui::dispatch_key now routes
   filter-mode keys to OverviewTab regardless of the visible tab,
   matching the pre-refactor behavior (filter input also still
   draws above the status bar on every tab when active).

Group expand/collapse (Space, Left/Right, l) stays Overview-only
since Details has no group rows. Enter stays Overview-only since
its meaning there (open Details) is a no-op when already on the
Details tab.
